:sparkles: (polybar) add autorandr module

+++ .config/polybar/launch.sh
@@ -1,14 +1,21 @@
 #!/bin/bash
-
+# check is running
+if [[ $$ -ne $(pgrep -fo "$0") ]];then
+  echo "Already running"
+  exit 1
+fi
 # Terminate already running bar instances
 killall -q polybar
 
 # Wait until the processes have been shut down
 while pgrep -u $UID -x polybar >/dev/null; do sleep 1; done
-profile=$(autorandr --current) 
+profile=$(autorandr --detected)
 if [[ "$profile" == "minipc_single_usbc" ]];then
   master="DisplayPort-0"
-  slave="" 
+  slave=""
+elif [[ "$profile" == "minipc_dual" ]];then
+  master="HDMI-A-0"
+  slave="DisplayPort-0"
 fi
 
 if [[ "$(hostname)" == "kamo3workmini" ]];then
